Set Lambda EventMapping on Dynamo stream to only retry 3 times (instead of 10000 times)
The max retries is a new feature as of Nov 2019, and CDK v 1.22 doesn't have it built into their DynamoEventSource construct. Fortunately, CDK has a first-class support for handling this situation: escape hatches. Unfortunately, The DynamoEventSource doesn't provide access to the underlying EventSourceMapping node so I can get to the underlying Cfn object. In fact, from looking at the code, it looks like they throw it away.

So this refactors our infra code to use a raw EventSourceMapping, and handle setting permissions ourselves.